refactor(models/stf): inline ELIC-style codec wiring; move convert to examples
Restyle WACNN / SymmetricalTransFormer after Elic2022Official:

- WACNN now subclasses SimpleVAECompressionModel (inherits the standard
  g_a -> latent_codec -> g_s forward/compress/decompress); STF keeps its
  CompressionModel base since its transformer transforms need height/width.
- The Family 1 entropy stack is wired inline in __init__ as two dict
  comprehensions (channel_context + per-slice latent_codec) feeding
  HyperpriorLatentCodec(... ChannelGroupsLatentCodec(...)), mirroring
  Elic2022Official channel_context / scctx_latent_codec layout. Drops the
  _build_family1_latent_codec + build_channel_slice_codec factory
  indirection.
- convert_upstream_stf_state_dict + its helpers (_nest_winmsa_keys,
  _is_upstream_stf_state_dict, _UPSTREAM_* constants) move to
  examples/convert_stf_checkpoint.py; from_state_dict now only loads
  already-converted (compressai-native) state dicts. State-dict paths are
  unchanged, so this is a pure code-organization change — round-trip
  diff stays 0.

tests: TestStf loads convert_upstream_stf_state_dict from the example
script via a new _load_convert_fn importlib helper (examples/ is not an
importable package).
